updated chatlog and filter rest api

+++ bansystem-extension-restapi/src/main/java/ch/dkrieger/bansystem/restapi/handler/defaults/FilterHandler.java
@@ -5,6 +5,7 @@ import ch.dkrieger.bansystem.lib.filter.Filter;
 import ch.dkrieger.bansystem.lib.filter.FilterOperation;
 import ch.dkrieger.bansystem.lib.filter.FilterType;
 import ch.dkrieger.bansystem.lib.utils.Document;
+import ch.dkrieger.bansystem.lib.utils.GeneralUtil;
 import ch.dkrieger.bansystem.restapi.ResponseCode;
 import ch.dkrieger.bansystem.restapi.handler.RestApiHandler;
 
@@ -16,19 +17,28 @@ public class FilterHandler extends RestApiHandler {
 
     @Override
     public void onRequest(Query query, Document response) {
-        if(query.contains("filter")){
+        if(query.contains("action")){
             if(query.get("action").equalsIgnoreCase("list")){
                 response.append("filters",BanSystem.getInstance().getFilterManager().getFilters());
                 return;
-            }else if(query.get("action").equalsIgnoreCase("create")){
+            }else if(query.get("action").equalsIgnoreCase("create") && query.contains("message")&& query.contains("type")){
                 FilterType type = FilterType.ParseNull(query.get("type"));
                 FilterOperation operation;
                 if(query.contains("operation")) operation = FilterOperation.ParseNull(query.get("operation"));
                 else operation = FilterOperation.CONTAINS;
+
+                if(type == null){
+                    response.append("code", ResponseCode.BAD_REQUEST).append("message","Invalid filter type");
+                    return;
+                }
+                if(operation == null){
+                    response.append("code", ResponseCode.BAD_REQUEST).append("message","Invalid operation type");
+                    return;
+                }
                 Filter filter = BanSystem.getInstance().getFilterManager().createFilterType(type,operation,query.get("message"));
                 response.append("filter",filter);
                 return;
-            }else if(query.get("action").equalsIgnoreCase("delete")){
+            }else if(query.get("action").equalsIgnoreCase("delete") && query.contains("id") && GeneralUtil.isNumber(query.get("id"))){
                 BanSystem.getInstance().getFilterManager().deleteFilter(Integer.valueOf(query.get("id")));
                 response.append("message","Filter deleted");
                 return;
